refactor: modals styles cleanup and subdirectories
