+++ webui/components/chat/attachments/inputPreview.html
@@ -1,6 +1,6 @@
 <script type="module">
     import { store } from "/components/chat/attachments/attachmentsStore.js";
-    import { store as imageViewerStore } from "/components/modals/image-viewer-store.js";
+    import { store as imageViewerStore } from "/components/modals/image-viewer/image-viewer-store.js";
 </script>
 
 <div x-data>
