+++ webui/components/chat/input/bottom-actions.html
@@ -2,9 +2,9 @@
 <head>
   <script type="module">
     import { store } from "/components/chat/input/input-store.js";
-    import { store as historyStore } from "/components/modals/history-store.js";
-    import { store as contextStore } from "/components/modals/context-store.js";
-    import { store as fileBrowserStore } from "/components/modals/file-browser-store.js";
+    import { store as historyStore } from "/components/modals/history/history-store.js";
+    import { store as contextStore } from "/components/modals/context/context-store.js";
+    import { store as fileBrowserStore } from "/components/modals/file-browser/file-browser-store.js";
   </script>
 </head>
 <body>
