+++ webui/components/chat/input/chat-bar-input.html
@@ -3,7 +3,7 @@
     <script type="module">
         import { store as speechStore } from "/components/chat/speech/speech-store.js";
         import { store as attachmentsStore } from "/components/chat/attachments/attachmentsStore.js";
-        import { store as fullScreenStore } from "/components/modals/full-screen-store.js";
+        import { store as fullScreenStore } from "/components/modals/full-screen-input/full-screen-store.js";
     </script>
 </head>
 <body>
@@ -50,6 +50,9 @@
             </button>
         </div>
     </div>
+
+    <!-- Full Screen Input Modal -->
+    <x-component path="modals/full-screen-input/full-screen-input.html"></x-component>
 
 
     <style>
