+++ webui/components/modals/context/context.html
@@ -2,7 +2,7 @@
 <head>
   <title>Context Window</title>
   <script type="module">
-    import { store } from "/components/modals/context-store.js";
+    import { store } from "/components/modals/context/context-store.js";
   </script>
 </head>
 <body>
