+++ webui/components/modals/file-browser/file-browser.html
@@ -2,7 +2,7 @@
 <head>
   <title>File Browser</title>
   <script type="module">
-    import { store } from "/components/modals/file-browser-store.js";
+    import { store } from "/components/modals/file-browser/file-browser-store.js";
   </script>
 </head>
 <body>
