+++ webui/components/modals/full-screen-input/full-screen-input.html
@@ -1,7 +1,7 @@
 <html>
 <head>
   <script type="module">
-    import { store } from "/components/modals/full-screen-store.js";
+    import { store } from "/components/modals/full-screen-input/full-screen-store.js";
   </script>
 </head>
 <body>
@@ -70,7 +70,83 @@
 
 
   <style>
-  /* No component-specific styles needed; styles are provided by css/modals.css and css/modals2.css */
+  /* Full Screen Input Modal Styles */
+  .full-screen-input-modal {
+    width: 90%;
+    max-width: 800px;
+    max-height: 80vh;
+    position: relative;
+    padding: 0;
+    background-color: rgb(20, 20, 20, 0.96);
+    border: 1.5px solid var(--color-border);
+  }
+  .full-screen-input-modal .modal-content {
+    height: calc(80vh);
+    padding: 0;
+    margin: 0;
+    overflow: hidden;
+  }
+  .full-screen-input-modal .modal-footer {
+    background: transparent;
+    max-height: 50px;
+  }
+  #full-screen-input {
+    width: 100%;
+    height: calc(100% - 50px);
+    border: none;
+    background-color: transparent;
+    color: var(--color-text);
+    font-family: "Roboto Mono", monospace;
+    font-optical-sizing: auto;
+    font-size: 0.955rem;
+    padding: 1.2rem 1rem;
+    resize: none;
+    outline: none;
+  }
+  .light-mode .full-screen-input-modal {
+    background-color: rgb(220, 220, 220, 0.86);
+  }
+  .full-screen-input-modal .modal-footer {
+    padding: 1rem 0;
+    border-top: none;
+    background: transparent;
+  }
+  .full-screen-input-modal h2 {
+    margin: 0;
+    padding: 0;
+    font-size: 1.1rem;
+    color: var(--color-text);
+    opacity: 0.8;
+  }
+  .full-screen-input-modal .modal-close {
+    position: absolute;
+    top: 1.2rem;
+    right: 1rem;
+    font-size: 1.5rem;
+    padding: 0 0.5rem;
+    line-height: 0.8;
+  }
+  .full-screen-input-modal .btn-ok {
+    margin-right: 1rem;
+  }
+  #full-screen-input::-webkit-scrollbar {
+    width: 6px;
+    height: 6px;
+  }
+  #full-screen-input::-webkit-scrollbar-track {
+    background: transparent;
+    margin: 14px;
+    border-radius: 6px;
+  }
+  #full-screen-input::-webkit-scrollbar-thumb {
+    background-color: rgba(155, 155, 155, 0.5);
+    border-radius: 6px;
+    -webkit-transition: background-color 0.2s ease;
+    transition: background-color 0.2s ease;
+  }
+  #full-screen-input::-webkit-scrollbar-thumb:hover {
+    background-color: rgba(155, 155, 155, 0.7);
+  }
   </style>
 </body>
 </html>
