+++ webui/components/modals/full-screen-input/full-screen-store.js
