+++ webui/components/modals/history/history.html
@@ -2,7 +2,7 @@
 <head>
   <title>History</title>
   <script type="module">
-    import { store } from "/components/modals/history-store.js";
+    import { store } from "/components/modals/history/history-store.js";
   </script>
 </head>
 <body>
