+++ webui/components/modals/image-viewer/image-viewer.html
@@ -4,7 +4,7 @@
     <title>Image Viewer</title>
 
     <script type="module">
-        import { store } from "/components/modals/image-viewer-store.js";
+        import { store } from "/components/modals/image-viewer/image-viewer-store.js";
     </script>
 </head>
