+++ webui/index.html
@@ -12,7 +12,6 @@
     <link rel="stylesheet" href="css/toast.css">
     <link rel="stylesheet" href="css/settings.css">
     <link rel="stylesheet" href="css/modals.css">
-    <link rel="stylesheet" href="css/modals2.css">
     <link rel="stylesheet" href="css/speech.css">
     <link rel="stylesheet" href="css/scheduler-datepicker.css">
     <link rel="stylesheet" href="css/notification.css">
@@ -1180,9 +1179,6 @@
             </div>
         </template>
     </div>
-
-    <!-- Full Screen Input Modal -->
-    <x-component path="modals/full-screen-input.html"></x-component>
 
     <!-- Drag and Drop Overlay Component -->
     <x-component path="chat/attachments/dragDropOverlay.html"></x-component>
